Monitor Spreadsheet
Created the Monitor spreadsheet which basically just took the working
Algo sheet and added parsing for the option computation numbers.  It
seems to basically work,  although it is very basic. It shows the Delta
by summing up the positions.  The gamma numbers do not work at the
moment.
The theta numbers seem to work but have not been vetted.