Improvements to Monitoring
This version of the sheet has the complete logic to apply the quadratic
roots to the equation of figuring out where the stops should go.
It also has the beginings of the watchlist functionality which is
designed to load the market data into the instrument and then display
that instead of writing the data directly to the sheet.  This is
required because many of the values do not actually come in properly all
the time.  By seperating it, this will allow the property functions to
filter out values that are bad.