AutoStopper is running
This version has the Autostopper running.   The remaining problem with
it, though, is that orders get lost in the display.  It would seem that
something errors out, and so the code path is not completing.  The
strategy to fix it is to clean up all of the items causing errors.