Autostopper is working
This version has the working auto-stopper in it.  There are a few issues
that would still need to be worked out, but in general it seems pretty
robust.  The issues I am aware of are that the order type is RTH only,
and as such the stops won't trigger in the morning.  Other than that, it
looks good.