Replacing Monitor
This commit really just goes back to an earlier version of Monitor.xlsm.
The newer versions ceased working.  This version works.  The item that
needs to be fixed now is that the AutoStopper is hyperactive.  For some
reason it does not respect the fact that there is a generous tick
tolerance specified.