Autostop Jittter Fixed
The problem with the autostop jitter was that the delta numbers being
fed in from IB were incorrect and changed frequently, and without merit
- that is the underlying barely moved but the changes in delta were very
large.
In order to fix this a few things happened.  First, the portfolio delat
and gamma are now written to the analytics file.  NB:  The location of
the values for portfolio delta and gamma are HARDCODED!!!!!!!!!! in the
Greeks module.  This will break!!!!!!!!!!

Next, there is a "filter" applied to the LET delta property of market
value.  There is a constant defined at the top of MarketData which
specifies the tolerance.  this is currently set to 1%.   Basically if
the delta of an option changes by more than this tolerance on one
update, the update is discarded and a message is written to the log.

Logging is now in place in the AutoStop where you cancel and replace an
order indicating why the order is being cancelled and replaced.