Added Status Line and Zero Qty order check
This commit includes a few features that were previously checked in
along the back-up branch.  They include the code for the order status
line, and the code which checks to make sure there is not a zero qty
order trying to be replaced.  It will fail anyway.
there is a definite difference of behavior when the "Bypass Precautions"
flag is set in the API settings.   when this flag is unset-meaning the
precautions are enabled, the order status messages don't seem to show
up, even though the order goes live.  I need to sort this one out.