Error 383 Handling
Error 383 is received when the system violates the API precautionary
settings.  It seems that the error message is sent but the order is also
placed.  This version handles that by routing to the order status
message with a status of "API Precaution".  This seems to work in terms
of not placing multiple errant orders, BUT the cancellation status is
not handled correctly.    The orders are cancelled in TWS but are not
reflected as such in the sheet