Changes to nextValidID and logging totalTimeValue
This version includes a change where the nextValidId number is created
within the createOrder sub.  This is instead of calling the API function
call.  This change was motivated by the Autostopper where two orders
were being placed back to back and the second one would fail because the
netValidId was not yet retrieved from the API.
This version also logs the totalTimeValue and allowableLoss.  Note that
it just reads them off of the front sheet.  This will break, but I
believe it is related to the instability in the autostop orders.