Filter changes to ImpliedVol
It appears that large changes to the implied vol calculation were
contributing to the portfolio being very twitchy.  I have implemented a
change to the let function for implied vol which discards changes
greater than a threshold number.  Currently this is set to 10% meaning
that if a quote comes in which moves the implied vol number by more than
10% (i.e. from 10 .0% to 11.1% in one step) this change is discarded.